requirements.txt
flask==2.3.3
tensorflow==2.18.1
mysql-connector-python==8.1.0
pillow>=10.2.0
numpy>=1.26.0,<2.1.0
scikit-learn>=1.4.0
requests==2.31.0
flask-cors
python-dotenv==1.0.0